+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.4.0] - 2018-02-03
+### Added
+Multiple requests concurrently for getting packages using promises and asynchronous requests
+
+### Changed
+Fixed service container name in commands
+
 ## [1.3.1] - 2017-12-30
 ### Changed
 Links on documentation at README.md